Set up process
This does a few things: first, we are set up to use the csv.gz zipped files to save some space. Second, I added a script to copy the files from my local box drive. Might replace this with a direct download link at some point, but for now, we'll roll with this.